More settings cleanup
Moves SettingsBase to Internal.ExceptionalSettingsBase to remove it from intellisense for users.
