+++ src/StackExchange.Exceptional.AspNetCore/Extensions.cs
@@ -38,7 +38,7 @@ namespace StackExchange.Exceptional
             Dictionary<string, string> customData = null,
             string applicationName = null)
         {
-            if (SettingsBase.IsLoggingEnabled)
+            if (ExceptionalSettingsBase.IsLoggingEnabled)
             {
                 try
                 {
@@ -87,7 +87,7 @@ namespace StackExchange.Exceptional
             Dictionary<string, string> customData = null,
             string applicationName = null)
         {
-            if (SettingsBase.IsLoggingEnabled)
+            if (ExceptionalSettingsBase.IsLoggingEnabled)
             {
                 try
                 {
