+++ src/StackExchange.Exceptional.Shared/Error.cs
@@ -32,7 +32,7 @@ namespace StackExchange.Exceptional
         /// Note: THIS IS LIKELY THE GLOBAL INSTANCE, so changes here likely impact all exceptions.
         /// </summary>
         [JsonIgnore]
-        public SettingsBase Settings { get; set; }
+        public ExceptionalSettingsBase Settings { get; set; }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="Error"/> class.
@@ -49,7 +49,7 @@ namespace StackExchange.Exceptional
         /// <param name="rollupPerServer">Whether to log up per-server, e.g. errors are only duplicates if they have same stack on the same machine.</param>
         /// <param name="initialCustomData">The initial data dictionary to start with (generated by the user).</param>
         public Error(Exception e,
-            SettingsBase settings,
+            ExceptionalSettingsBase settings,
             string category = null,
             string applicationName = null,
             bool rollupPerServer = false,
@@ -175,7 +175,7 @@ namespace StackExchange.Exceptional
         }
 
         /// <summary>
-        /// Populates the CustomData collection via <see cref="SettingsBase.GetCustomData"/>, if set.
+        /// Populates the CustomData collection via <see cref="ExceptionalSettingsBase.GetCustomData"/>, if set.
         /// </summary>
         private void AddCustomData()
         {
