+++ src/StackExchange.Exceptional.Shared/Internal/ExceptionalSettingsBase.cs
@@ -1,18 +1,15 @@
-﻿using StackExchange.Exceptional.Internal;
-using StackExchange.Exceptional.Notifiers;
+﻿using StackExchange.Exceptional.Notifiers;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Net;
-using System.Net.Mail;
 using System.Text.RegularExpressions;
 
-namespace StackExchange.Exceptional
+namespace StackExchange.Exceptional.Internal
 {
     /// <summary>
     /// Settings for Exceptional error logging.
     /// </summary>
-    public abstract class SettingsBase
+    public abstract class ExceptionalSettingsBase
     {
         /// <summary>
         /// Event handler to run before an exception is logged to the store.
@@ -208,206 +205,5 @@ namespace StackExchange.Exceptional
         /// Settings for prettifying a StackTrace
         /// </summary>
         public StackTraceSettings StackTrace { get; } = new StackTraceSettings();
-    }
-
-    /// <summary>
-    /// A settings object got setting up an error store.
-    /// </summary>
-    public class ErrorStoreSettings
-    {
-        /// <summary>
-        /// Application name to log with.
-        /// </summary>
-        public string ApplicationName { get; set; } = "My Application";
-
-        /// <summary>
-        /// The type of error store to use, File, SQL, Memory, etc.
-        /// </summary>
-        public string Type { get; set; }
-
-        /// <summary>
-        /// For file-based error stores.
-        /// The path to use on for file storage.
-        /// </summary>
-        public string Path { get; set; }
-
-        /// <summary>
-        /// For database-based error stores.
-        /// The connection string to use.  If provided, ConnectionStringName is ignored.
-        /// </summary>
-        public string ConnectionString { get; set; }
-
-#if !NETSTANDARD2_0
-        /// <summary>
-        /// For database-based error stores.
-        /// The name of the connection string to use from the application's configuration.
-        /// </summary>
-        public string ConnectionStringName { get; set; }
-#endif
-
-        /// <summary>
-        /// The size of this error log, either how many to keep or how many to display depending on type.
-        /// Defaults to 200.
-        /// </summary>
-        public int Size { get; set; } = 200;
-
-        /// <summary>
-        /// The duration of error groups to roll-up, similar errors within this timespan (those with the same stack trace) will be shown as duplicates.
-        /// Defaults to 10 minutes.
-        /// </summary>
-        public TimeSpan? RollupPeriod { get; set; } = TimeSpan.FromMinutes(10);
-
-        /// <summary>
-        /// The size of the backup queue to use for the log, after roll-ups, it's how many entries in memory can be stored before culling the oldest.
-        /// Defaults to 1000.
-        /// </summary>
-        public int BackupQueueSize { get; set; } = 1000;
-
-        /// <summary>
-        /// When a connection to the error store failed, how often to retry logging the errors in queue for logging.
-        /// </summary>
-        public TimeSpan BackupQueueRetryInterval { get; set; } = TimeSpan.FromSeconds(2);
-    }
-
-    /// <summary>
-    /// Email settings configuration, for configuring Email sending from code.
-    /// </summary>
-    public class EmailSettings
-    {
-        private string _fromAddress, _fromDisplayName,
-                       _SMTPUserName, _SMTPPassword;
-
-        internal MailAddress FromMailAddress { get; private set; }
-        internal NetworkCredential SMTPCredentials { get; private set; }
-
-        private void SetMailAddress()
-        {
-            try
-            {
-                // Because MailAddress.TryParse() isn't a thing, and an invalid address will throw.
-                FromMailAddress = _fromDisplayName.HasValue()
-                                  ? new MailAddress(_fromAddress, _fromDisplayName)
-                                  : new MailAddress(_fromAddress);
-            }
-            catch
-            {
-                FromMailAddress = null;
-            }
-        }
-
-        private void SetCredentials() =>
-            SMTPCredentials = _SMTPUserName.HasValue() && _SMTPPassword.HasValue()
-                              ? new NetworkCredential(_SMTPUserName, _SMTPPassword)
-                              : null;
-
-        /// <summary>
-        /// The address to send email messages to.
-        /// </summary>
-        public string ToAddress { get; set; }
-        /// <summary>
-        /// The address to send email messages from.
-        /// </summary>
-        public string FromAddress
-        {
-            get => _fromAddress;
-            set
-            {
-                _fromAddress = value;
-                SetMailAddress();
-            }
-        }
-        /// <summary>
-        /// The display name to send email messages from.
-        /// </summary>
-        public string FromDisplayName
-        {
-            get => _fromDisplayName;
-            set
-            {
-                _fromDisplayName = value;
-                SetMailAddress();
-            }
-        }
-        /// <summary>
-        /// The SMTP server to send mail through.
-        /// </summary>
-        public string SMTPHost { get; set; }
-        /// <summary>
-        /// The port to send mail on (if SMTP server is specified via <see cref="SMTPHost"/>).
-        /// </summary>
-        public int? SMTPPort { get; set; }
-        /// <summary>
-        /// The SMTP user name to use, if authentication is needed.
-        /// </summary>
-        public string SMTPUserName
-        {
-            get => _SMTPUserName;
-            set
-            {
-                _SMTPUserName = value;
-                SetCredentials();
-            }
-        }
-        /// <summary>
-        /// The SMTP password to use, if authentication is needed.
-        /// </summary>
-        public string SMTPPassword
-        {
-            get => _SMTPPassword;
-            set
-            {
-                _SMTPPassword = value;
-                SetCredentials();
-            }
-        }
-        /// <summary>
-        /// Whether to use SSL when sending via SMTP.
-        /// </summary>
-        public bool SMTPEnableSSL { get; set; }
-        /// <summary>
-        /// Flags whether or not emails are sent for duplicate errors.
-        /// </summary>
-        public bool PreventDuplicates { get; set; }
-    }
-
-    /// <summary>
-    /// Settings for prettifying a StackTrace
-    /// </summary>
-    public class StackTraceSettings
-    {
-        /// <summary>
-        /// Replaces generic names like Dictionary`2 with Dictionary&lt;TKey,TValue&gt;.
-        /// Specific formatting is based on the <see cref="Language"/> setting.
-        /// </summary>
-        public bool EnablePrettyGenerics { get; set; } = true;
-        /// <summary>
-        /// The language to use when prettifying StackTrace generics.
-        /// Defaults to C#.
-        /// </summary>
-        public CodeLanguage Language { get; set; }
-        /// <summary>
-        /// Whether to print generic type names like &lt;T1, T2&gt; etc. or just use commas, e.g. &lt;,,&gt; if <see cref="Language"/> is C#.
-        /// Defaults to true.
-        /// </summary>
-        public bool IncludeGenericTypeNames { get; set; } = true;
-
-        /// <summary>
-        /// The language to use when operating on errors and stack traces.
-        /// </summary>
-        public enum CodeLanguage
-        {
-            /// <summary>
-            /// C#
-            /// </summary>
-            CSharp,
-            /// <summary>
-            /// F#
-            /// </summary>
-            FSharp,
-            /// <summary>
-            /// Visual Basic
-            /// </summary>
-            VB
-        }
     }
 }
