+++ src/StackExchange.Exceptional/Extensions.cs
@@ -35,7 +35,7 @@ namespace StackExchange.Exceptional
             Dictionary<string, string> customData = null,
             string applicationName = null)
         {
-            if (SettingsBase.IsLoggingEnabled)
+            if (ExceptionalSettingsBase.IsLoggingEnabled)
             {
                 try
                 {
@@ -84,7 +84,7 @@ namespace StackExchange.Exceptional
             Dictionary<string, string> customData = null,
             string applicationName = null)
         {
-            if (SettingsBase.IsLoggingEnabled)
+            if (ExceptionalSettingsBase.IsLoggingEnabled)
             {
                 try
                 {
